+++ Src/LibraryCore.CommandLineParser/Runner.cs
@@ -60,13 +60,4 @@ public static class Runner
 
         return parameters;
     }
-
-    public record InvokeParameters
-    {
-        public IImmutableList<CommandConfiguration> ConfiguredCommands { get; init; } = null!;
-        public IDictionary<string, string> RequiredParameters { private get; init; } = null!;
-        public Action<string> MessagePump { get; init; } = null!;
-
-        public T RequiredParameterToValue<T>(string key) => (T)Convert.ChangeType(RequiredParameters[key], typeof(T));
-    }
 }
